feat: change error messages

+++ src/posts/posts.service.ts
@@ -29,7 +29,7 @@ export class PostsService {
 
   async findOne(id: string): Promise<Post> {
     if (!Types.ObjectId.isValid(id)) {
-      throw new BadRequestException('Invalid post ID format');
+      throw new BadRequestException('Invalid ID');
     }
 
     const post = await this.postModel
@@ -37,14 +37,19 @@ export class PostsService {
       .populate('author', 'username')
       .exec();
     if (!post) {
-      throw new NotFoundException(`Post with ID ${id} not found`);
+      throw new NotFoundException('Invalid ID');
     }
     return post;
   }
 
   async delete(id: string): Promise<any> {
     if (!Types.ObjectId.isValid(id)) {
-      throw new BadRequestException('Invalid post ID format');
+      throw new BadRequestException('Invalid ID');
+    }
+
+    const post = await this.postModel.findById(id).exec();
+    if (!post) {
+      throw new NotFoundException('Invalid ID');
     }
 
     return this.postModel.findByIdAndDelete(id).exec();
@@ -52,14 +57,14 @@ export class PostsService {
 
   async update(id: string, updatePostDto: Partial<Post>): Promise<Post> {
     if (!Types.ObjectId.isValid(id)) {
-      throw new BadRequestException('Invalid post ID format');
+      throw new BadRequestException('Invalid ID');
     }
 
     const updatedPost = await this.postModel
       .findByIdAndUpdate(id, updatePostDto, { new: true })
       .exec();
     if (!updatedPost) {
-      throw new NotFoundException(`Post with ID ${id} not found`);
+      throw new NotFoundException('Invalid ID');
     }
 
     return updatedPost.populate('author', 'username');
